requirements.txt
huggingface_hub==0.25.2
gradio>=4.0.0
langchain>=0.1.0
langchain-community>=0.0.11
langchain-core>=0.1.10
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0